+++ kashikishi/src/main.rs
@@ -27,7 +27,7 @@ use font_rasterizer::{
     layout_engine::{Model, ModelOperation, World},
     rasterizer_pipeline::Quarity,
     support::{
-        action_processor::{ActionProcessor, ActionProcessorStore},
+        action::{ActionProcessor, ActionProcessorStore},
         run_support, Flags, InputResult, SimpleStateCallback, SimpleStateSupport,
     },
     time::set_clock_mode,
